pyproject.toml
[tool.poetry]
name = "iskra"
version = "1.0.0"
description = "extract statistics from ISKRA smart meter"
authors = ["Alexander Dreweke <thrawn@shadowhunt.de>"]
license = "GPL-3.0-or-later"
readme = "README.md"

[tool.poetry.dependencies]
argparse = "^1.4.0"
psycopg = "^3.1.1"
python = "^3.9"
smllib = "^1.2"

[tool.poetry.dev-dependencies]
flake9 = "^3.8.3"
pyupgrade = "^2.37.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
